Update script by comments
